fix(specification): fix specification invalid type data

+++ src/main/java/com/chris/cmarket/Specifications/Globals/PriceSpecification.java
@@ -1,5 +1,7 @@
 package com.chris.cmarket.Specifications.Globals;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.jpa.domain.Specification;
 
 /**
@@ -17,7 +19,7 @@ public class PriceSpecification<T> {
      * @param <T> The entity type.
      * @return The specification or null if min is null.
      */
-    public static <T> Specification<T> withMin(Double min) {
+    public static <T> Specification<T> withMin(BigDecimal min) {
         return (root, query, cb) -> min == null ? null : cb.greaterThanOrEqualTo(root.get("price"), min);
     }
 
@@ -29,7 +31,7 @@ public class PriceSpecification<T> {
      * @param <T> The entity type.
      * @return The specification or null if max is null.
      */
-    public static <T> Specification<T> withMax(Double max) {
+    public static <T> Specification<T> withMax(BigDecimal max) {
         return (root, query, cb) -> max == null ? null : cb.lessThanOrEqualTo(root.get("price"), max);
     }
 }
